logout 2021/02/08

+++ autoTA/users/views.py
@@ -7,42 +7,55 @@ import json
 from django.core.serializers.json import DjangoJSONEncoder
 
 def add_favorite(request, stock_code):
-    data = json.loads(request.body)
-    user_email = data['token']
-    if auth.valid_user(user_email):
-        stocks_api.add_favorite(user_email,stock_code)
-        favorites = stocks_api.get_favorites(user_email)
-        favorites = json.dumps(favorites, cls=DjangoJSONEncoder,ensure_ascii = False)
-        return HttpResponse(favorites)
+    data = request.COOKIES
+    if 'token' in data:
+        user_email = data['token']
+        if auth.valid_user(user_email):
+            stocks_api.add_favorite(user_email,stock_code)
+            favorites = stocks_api.get_favorites(user_email)
+            favorites = json.dumps(favorites, cls=DjangoJSONEncoder,ensure_ascii = False)
+            return HttpResponse(favorites)
+    success = {'success':False, 'message':"no user"}
+    return HttpResponse(json.dumps(success))
 
 def delete_favorite(request, stock_code):
-    data = json.loads(request.body)
-    user_email = data['token']
-    if auth.valid_user(user_email):
-        stocks_api.delete_favorite(user_email,stock_code)
-        favorites = stocks_api.get_favorites(user_email)
-        favorites = json.dumps(favorites, cls=DjangoJSONEncoder,ensure_ascii = False)
-        return HttpResponse(favorites)
+    data = request.COOKIES
+    if 'token' in data:
+        user_email = data['token']
+        if auth.valid_user(user_email):
+            stocks_api.delete_favorite(user_email,stock_code)
+            favorites = stocks_api.get_favorites(user_email)
+            favorites = json.dumps(favorites, cls=DjangoJSONEncoder,ensure_ascii = False)
+            return HttpResponse(favorites)
+    success = {'success':False, 'message':"no user"}
+    return HttpResponse(json.dumps(success))
 
 def get_favorites(request):
-    data = json.loads(request.body)
-    user_email = data['token']
-    if auth.valid_user(user_email):
-        favorites = stocks_api.get_favorites(user_email)
-        favorites = json.dumps(favorites, cls=DjangoJSONEncoder,ensure_ascii = False)
-        return HttpResponse(favorites)
+    data = request.COOKIES
+    if 'token' in data:
+        user_email = data['token']
+        if auth.valid_user(user_email):
+            favorites = stocks_api.get_favorites(user_email)
+            favorites = json.dumps(favorites, cls=DjangoJSONEncoder,ensure_ascii = False)
+            return HttpResponse(favorites)
+    success = {'success':False, 'message':"no user"}
+    return HttpResponse(json.dumps(success))
 
 def get_user(request):
-    data = json.loads(request.COOKIE)
-    valid = auth.valid_user(data['token'])
-    if(valid):
-        user = auth.get_user_info(data['token'])
-        success = {'success':valid, 'token':user}
-        response = HttpResponse(json.dumps(success))
-        response.set_cookie('token',user)
-        return response
+    data = request.COOKIES
+    if 'token' in data:
+        valid = auth.valid_user(data['token'])
+        if(valid):
+            user = auth.get_user_info(data['token'])
+            success = {'success':valid, 'token':user.email}
+            response = HttpResponse(json.dumps(success))
+            response.set_cookie('token',user)
+            return response
+        else:
+            success = {'success':valid}
+            return HttpResponse(json.dumps(success))
     else:
-        success = {'success':valid}
+        success = {'success':False}
         return HttpResponse(json.dumps(success))
 
 def register(request):
@@ -55,5 +68,7 @@ def register(request):
     return response
 
 def logout(request):
+    success = {'success':True, 'message':'logout success'}
+    response = HttpResponse(json.dumps(success))
     response.delete_cookie('token')
     return response
